+++ src/components/Example/Example.js
@@ -27,39 +27,46 @@ export default class Example extends Component {
       example = examples[0]
     }
 
-    const isJson = typeof example !== 'string'
-
     if (!example) {
       return null
     }
 
+    const isJson = typeof example.value !== 'string'
+
     return (
-      <div className={classes.example}
-        onMouseEnter={() => this.setState({ hovered: true })}
-        onMouseLeave={() => this.setState({ hovered: false })}
-      >
-        <div className={classNames(classes.buttons, {
-          [classes.hovered]: this.state.hovered
-        })}>
-          <CopyButton onCopyClick={this.onCopyClick} tooltip='Copy to Clipboard' />
-          {!isJson &&
-            <span onClick={() => this.setState({ collapseAll: false })}>Expand All</span>}
-          {!isJson &&
-            <span onClick={() => this.setState({ collapseAll: true })}>Collapse All</span>}
-        </div>
-        <pre>
-          {
-            isJson
-              ? JSON.stringify(example, null, 2)
-              : example
-          }
-        </pre>
+      <div className={classes.example}>
+        {example.summary &&
+        <div className={classes.summary}>{ example.summary }</div>}
+        {example.description &&
+        <div className={classes.description}>{ example.description }</div>}
+        {example.value &&
+        <div className={classes.value}
+          onMouseEnter={() => this.setState({ hovered: true })}
+          onMouseLeave={() => this.setState({ hovered: false })}
+        >
+          <div className={classNames(classes.buttons, {
+            [classes.hovered]: this.state.hovered
+          })}>
+            <CopyButton onCopyClick={this.onCopyClick} tooltip='Copy to Clipboard' />
+            {!isJson &&
+            <span onClick={() => this.setState({collapseAll: false})}>Expand All</span>}
+            {!isJson &&
+            <span onClick={() => this.setState({collapseAll: true})}>Collapse All</span>}
+          </div>
+          <pre>
+            {
+              isJson
+                ? JSON.stringify(example.value, null, 2)
+                : example.value
+            }
+          </pre>
+        </div>}
       </div>
     )
   }
 
   onCopyClick () {
-    copy(JSON.stringify(this.props.examples[0], null, 2))
+    copy(JSON.stringify(this.props.examples[0].value, null, 2))
   }
 }