+++ CHANGELOG.md
@@ -4,11 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [Unreleased][]
 
-Nil.
+## [0.3.0][] - 2017-10-24
 
-## [v0.2.0] - 2017-09-14
+### Added
+
+- [#199](https://github.com/temando/open-api-renderer/issues/199) `navigationMethodDisplayType` parameter is added for better customisation of the navigation on the left.
+
+## [0.2.1][] - 2017-10-18
+
+### Added
+
+- Updated example object rendering support the summary, description, and value attributes.
+- Implement CHANGELOG handling with `changelog-verify` and `version-changelog`.
+
+### Fixed
+
+- [#196](https://github.com/temando/open-api-renderer/issues/196) Validation changed to use POST instead of GET
+
+## [0.2.0][] - 2017-09-14
 
 ### Changed
 
@@ -20,10 +35,7 @@ Nil.
 - [#192](https://github.com/temando/open-api-renderer/issues/192) Swap `isomorphic-fetch` for `fetch-everywhere` to prevent `ReferenceError: self is not defined` error.
 - [#176](https://github.com/temando/open-api-renderer/issues/176) Lincoln learnt how to handle circular reference in schema.
 
-### Added
-- [#115](https://github.com/temando/open-api-renderer/issues/115) Visualise per method the security schemes required.
-
-## [v0.1.0] - 2017-07-17
+## [0.1.0][] - 2017-07-17
 
 ### Added
 
@@ -31,6 +43,13 @@ Nil.
 - [#157](https://github.com/temando/open-api-renderer/issues/157) Lincoln learnt how to display `servers` and path information.
 - [#112](https://github.com/temando/open-api-renderer/issues/112) Lincoln learnt how to display `externalDocs`.
 
-## [v0.0.6] - 2017-07-07
+## [0.0.6][] - 2017-07-07
 
 - Initial release of Lincoln 🎩
+
+[Unreleased]: https://github.com/temando/open-api-renderer/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/temando/open-api-renderer/compare/v0.2.1...v0.3.0
+[0.2.1]: https://github.com/temando/open-api-renderer/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/temando/open-api-renderer/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/temando/open-api-renderer/compare/v0.0.6...v0.1.0
+[0.0.6]: https://github.com/temando/open-api-renderer/tree/v0.0.6